[TX sync] Exploratory fix for Tx sync issues (still under investigation)
 * Rudimentary fix for slow Tx sync ultimately resulting to incompletely
   synchronised Tx store (missing transactions)

+++ libs/ledger/include/ledger/storage_unit/transaction_store_sync_service.hpp
@@ -163,9 +163,9 @@ private:
   RequestingSubTreeList pending_subtree_;
   RequestingTxList      pending_objects_;
 
-  std::queue<uint8_t>                                          roots_to_sync_;
+  std::queue<uint64_t>                                         roots_to_sync_;
   uint64_t                                                     root_size_ = 0;
-  std::unordered_map<PromiseOfTxList::PromiseCounter, uint8_t> promise_id_to_roots_;
+  std::unordered_map<PromiseOfTxList::PromiseCounter, uint64_t> promise_id_to_roots_;
 
   std::atomic_bool is_ready_{false};
 };
